Fix alarms issues on non-UTC systems
This fixes two issues
-Alarm clear time incorrectly reported on non-UTC systems.
-Paused instances do not display the correct alarm timestamp in
non-UTC systems.

The mktime call that's made in fm_db_util_get_timestamp uses the local
tz in it's calculation. It assumes the incoming time is also in the
local time. When mktime is provided with a value in UTC we see the error
presented in this bug.

Change-Id: Iaf8d097e2dfdfae585b6976e02a85b9e2f1b19f9
Closes-Bug: 1828448
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>