Set management affecting severity for specific app alarms
In this commit, the Management_Affecting_Severity is
changed from none to warning for 2 types of k8s app alarms:
  - Application Apply In Progress (alarm id: 750.004) and
  - Application Update In Progress (alarm id: 750.005)

This change will disallow a patch or upgrade while any
of these alarms exists.

The Degrade_Affecting_Severity setting will remain to
be 'none' for all app alarms as logically these events
should not degrade the host to the point of disabling
filesystem resizing or controller swact.

Closes-Bug: 1833323
Change-Id: Ie5a85bb4480f8ce6dcb63dfd2e26dacdbceeb366
Signed-off-by: Tee Ngo <Tee.Ngo@windriver.com>