Fix py27 and py35 zuul jobs failing to setup
As of March 1, when jobs run on ubuntu-xenial, they fail
to run py27 and py35 jobs with an error similar to:

/usr/bin/python3 -m virtualenv --no-download
  --python /usr/bin/python2.7 py27

ImportError: cannot import name 'ContextManager'

It does not appear to impact the ubuntu-bionic nodes so
updating zuul to use those nodes.

bionic nodes do not support python3.5, so the py35 jobs
is now converted to py36.

Partial-Bug: 1865554
Change-Id: Iad13d03ced21294d1a02c82f4e1e37cac3f27ee7
Signed-off-by: albailey <Al.Bailey@windriver.com>