Fix zuul pylint failure on review
For a review in stx/fault repo if pylint runs on ubuntu-focal,
it'll fail with error "ERROR: InvocationError for command ...".
This update fixed this by restricting pylint to run on
the compatible ubuntu-bionic, as we do for stx/config repo.

Closes-Bug: 1930919
Signed-off-by: Andy Ning <andy.ning@windriver.com>
Change-Id: I584e80d6286dbaf5439cbff593d0ad74da96ab77
(cherry picked from commit 60f7cb230899f6960d57394453514a57690414ca)