Re-enable python2 support
In order for the centos8 branch to merge back into master, it has
to work with both python2 and python3. However, in a previous commit
python2 support was removed in fm-common.

To support both python2 and python3, first we detect whether the base
OS is either running Centos7 or Centos8 and select the correct LDFLAGS
and use the correct CPPFLAGS. At runtime we determine the python version
and use the preprocessor to do the right thing.

Story: 2006796
Task: 42813

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I3d83aaf8086e780603f9a227a4baad2d4e519abd