Fix quotes in new alarm doc
Recent commit to add Expiring-Soon and Expired alarm
IDs and description resulted in failure in Openstack
app with following error.

UnicodeEncodeError: 'ascii' codec can't encode character
u'\u2018' in position 12: ordinal not in range(128)

The problem was traced to incorrect quotation marks
in Description field. Tested with a fix of the quotes
and fm-rest-api pod was able to come up successfully.

Story: 2008946
Task: 43049

Signed-off-by: Sabeel Ansari <Sabeel.Ansari@windriver.com>
Change-Id: Iacdf8c18daf412b0ac6d9d8aa89de1c4026356b7