debian: Fix python3 syntax warning
From https://docs.python.org/3/whatsnew/3.8.html#changes-in-python-behavior:

The compiler now produces a SyntaxWarning when identity
checks (is and is not) are used with certain types of
literals (e.g. strings, numbers). These can often work
by accident in CPython, but are not guaranteed by the
language spec. The warning advises users to use
equality tests (== and !=) instead.

Test Plan:
PASS Apply patch test the output of fm alarm-list on Debian 11.
PASS Apply patch test the output of fm alarm-list on Centos 7.

Closes-Bug: 1960958

Signed-off-by: Chuck Short <charles.short@windriver.com>
Change-Id: I12c43f336a449876fab8592e34eace674518d0da