Update events.yaml for specific alarms
The following alarms need to be clearer for users. Some information
was gathered in order to improve the 'description' and 'Proposed
repair action' fields.

Alarm that has been deprecated:
270.001

The list of alarms that were modified is the following:
500.210
500.200
750.002
750.006
800.003
800.102
900.002
900.003
900.004
900.009
900.103
900.203
900.303
900.503

Test plan:
There is no need to test the alarms affected by the changes.

Story: 2010143
Task: 45785

Signed-off-by: Agustin Carranza <agustin.carranza@windriver.com>
Change-Id: I57b86548a36da66119cb04779ce1f9147254316c