Extend events.yaml schema with usage context field
The events.yaml file contains every alarm and log used by platform and
openstack. There is no way to know which one relates to one or
the other.
In order to know that, it is required an additional field as part of
each record to differential between platform and openstack.

Story: 2010143
Task: 46723

Test plan
PASS: Build the fm-api and fm-doc packages.
      Install fm-api first and then fm-doc.
      No errors are found during build and installation process.

Signed-off-by: Agustin Carranza <agustin.carranza@windriver.com>
Change-Id: I8598afc77d27d107c4f9a108dd46b2ebc79b30a1