Add ceph commands in the 800 series alarm document
When a 800-Series alarm occurs, users refer to the documentation to
know what kind of error is shown. But sometimes that is not enough
information.
The output of some commands can be useful information and could
save time when solving issues related to the storage alarms.

Closes-bug: 2004601

Test plan
PASS: * Build fm packages and deploy an ISO containing new fm
        packages.
      * Trigger alarms that were modified by this commit,
        (e.g. shutdown a controller).
      * Run fm alarm-list --uuid and copy the uuid of a 800-series
        alarm.
      * Run fm alarm-show <uuid> and check that the field
        has changed.

Signed-off-by: Agustin Carranza <agustin.carranza@windriver.com>
Change-Id: I94e2719b55b4fc14b692439526b5b47204460ac7