Fix Context value for some alarms
Some documentation generating scripts were introduced in order to avoid
manual intervention every time an alarm/log is changed/added/removed.
Those scripts required a way to know where the alarm/log belongs to.
For that requirement, the field Context was introduced in previous
commits. During that development, it was taken the current
classification at that time in the docs as source of truth, but it was
outdated.

This commits modifies the values that were detected as wrong/outdated.
The scripts also require the value 'none' in the Context field for when
an alarm/log should not be included in the documentation but still be
defined in the events.yaml file. So the Context value is updated for
that case too.

Context incorrectly tagged as openstack and changed to starlingx:
* 900.006

Context incorrectly tagged as starlingx and changed to openstack:
* 100.105
* 100.112
* 100.113
* 300.001
* 300.002

Closes-bug: 2012981

Test plan
PASS: Since the Context field does not have impact in functionality,
      build and install fm-doc package successfully.
      Check the file in the filesystem contains this change.
PASS: Trigger random alarms and check FM functionality.

Signed-off-by: Agustin Carranza <agustin.carranza@windriver.com>
Change-Id: I16f858bbb712349f08b2ceca33152e365b0ed733