Add parsing check when Context field is empty
This change adds a parsing check to ERROR if Context field is Empty.
Until now there had not been a requirement of non empty fields, so in
case this is needed in the future for other key/values, a collection is
created.

Test plan
PASS: * Add/modify an alarm/log in events.yaml file with Context field
        set to <Empty>.
      * Run the checkEventYaml script and check it fails.
PASS: * Check that all the events in events.yaml file have the Context
        field set to a non empty value.
      * Run the checkEventYaml script and check it ends successfully.

Closes-bug: 2020381

Signed-off-by: Agustin Carranza <agustin.carranza@windriver.com>
Change-Id: Ia267886dd49099525751165975fb5d291c0c6f82