Link alarms to current documentation URL
Since the alarm documentation has been automated and the events.yaml
file is taken as source of truth for it, it is required to link
the alarms proposed repair action with a direct link to the
documentation for the users.

This change modifies the mentions of documentation to a proper link,
using Sphinx placeholder that are interpreted by the documentation
language.

Test plan
PASS:
* Build fm-doc package. Check that all parsing checks were run and
package was built successfully.

Closes-bug: 2022104

Change-Id: Iccb34e42ed80634d73cf7549e9230976579deef7
Signed-off-by: Agustin Carranza <agustin.carranza@windriver.com>