Add alarm id, definition for system peer and subcloud peer group
The alarm 280.004 is added and will be raised when the system peer
connection failure has been detected, and will be cleared when the
system peer connection has been restored.
The alarm 280.005 is added and will be raised when the a subcloud peer
group is being managed by a remote system with a lower priority, and
will be cleared when the subclouds belonging to the subcloud peer group
have been migrated back to current system.

Test Plan:
 PASS - Verify successful tox test and package build

Story: 2010852
Task: 48492
Change-Id: I3068676933c0446a88bd4290277456cd0962f941
Signed-off-by: Zhang Rong(Jon) <rong.zhang@windriver.com>