Add "proposed_repair_action" attr to event log file
Even when event logs have this attribute available, it is not being
logged in the /var/log/fm-event.log file. This could result in a lack
of certain information for external tools that use it.

Test Plan:
PASS: Build fm-common package and install it. Then, trigger an alarm
      with 'proposed_repair_action'. Verify its presence in the
      /var/log/fm-event.log
PASS: Raise an alarm without 'proposed_repair_action'. Check that
      the empty attribute is present.
PASS: Raise an alarm with a long 'proposed_repair_action'. verify that
      this message is not affecting other attributes.

Closes-Bug: 2042579

Change-Id: Ic27b840041872c3afd0be28e11556acf42a3d5a9
Signed-off-by: fperez <fabrizio.perez@windriver.com>