Add controller-fs alarm to fault management
This update introduces a new storage alarm: 800.105

The new alarm is related to the execution failure of an optional
controllerfs runtime manifest. The runtime manifest is executed
in association with creation/deletion of an optional controllerfs
filesystem. The alarms is host based (controllers) and is cleared
once the manifest executes successfully on that controller.

Test Plan:
 PASS: Build all FM packages and a new image with all the
       topic changes.
 PASS: AIO-DX Fresh install using the built image.
 PASS: Check that the alarm is raised if the creation/deletion
       of controller-fs fails.
 PASS: Check that the alarm is cleared when doing a successful
       retry.

Story: 2011117
Task: 50219

Change-Id: Iacd5785c033686b779670ca0ce3e2a38c921497f
Signed-off-by: Gabriel de Araújo Cabral <gabriel.cabral@windriver.com>