Update 700.016 alarm context to starlingx
Alarm 700.016 is raised and cleared by VIM.
This alarm is part of WRCP alarm list. So,
the alarm context is changed to 'starlingx'

Test Plan:

This is a document change. Changing the 'context'
to 'starlingx' will help pull the updates
automatically.

To raise alarm 700.016:
Controller's state is changed (either via swact or reboot)
such that it becomes the active controller and the uptime of
the controller is <20mins

Closes-Bug: 2078963
Change-Id: I28e1fc510d7839d13580124401db75ca07ef3fb4
Signed-off-by: Vanathi.Selvaraju <vanathi.selvaraju@windriver.com>