Prepare FM queries for sqlalchemy 1.4 upversion
For subcloud scalability purposes, sqlalchemy will be upgraded from
version 1.3 to 1.4. This version change introduces some
incompatibilities from the current version:

- Session management: When calling model_query without providing a
session, a new one would be created inside a context manager. This
means that, upon exiting model_query, the query would be executed and
any new filtering after would reopen the session. Version 1.3 was more
forgiving regarding transactions and this was not an issue.
On version 1.4, sessions called after transaction ended will still
remain open [1], causing the pool to eventually overflow with sessions
stuck in "idle in transaction" state. To fix this, the session
creation inside model_query is changed to only start the context
manager, which will then be closed only when exiting the db function
itself by a class decorator.

- Row values: Since there was a change in how row values are
accessed [2], added columns from different tables are not added
directly to the main result like before. Instead, the return value is
a tuple with the first element being the actual model and the
subsequent elements being any added columns. For this, attributes are
added to the original model with the extra columns names and values.

Test plan:
  - PASS: Build a custom ISO with the changes and deploy a DX system
          controller and a SX subcloud. Verify the system works as
          expected.
  - PASS: Manage a subcloud and verify the sync_status is "in-sync".
  - PASS: Soak the system and verify there was no connection leak and
          no sessions stuck in "idle in transaction" state.
  - PASS: Run DC sanity and regression.
  - PASS: Run the following commands and verify they work as expected
          and no errors were observed in the log file:
    - fm alarm-list
    - fm alarm-show
    - fm alarm-delete
    - fm event-list
    - fm event-show
    - fm event-suppress
    - fm event-unsupress
    - fm event-suppress-list
    - fm event-unsuppress-all

[1]: https://docs.sqlalchemy.org/en/14/changelog/migration_14.html#session-features-new-autobegin-behavior
[2]: https://docs.sqlalchemy.org/en/14/changelog/migration_14.html#rowproxy-is-no-longer-a-proxy-is-now-called-row-and-behaves-like-an-enhanced-named-tuple

Story: 2011311
Task: 51733

Change-Id: I44149317d6ce9e8ebfc851c2c3b2d66ff88166cb
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>