Using cache to improve performance in CLI context
In the first execution, the standard flow is performed by authenticating
the user with their credentials and the FM endpoint is discovered.
Then, the authentication token and service endpoint are stored in user's
session keyring and any future execution will reuse this information,
avoiding these steps.

When the token expires, the standard flow will be executed again,
restarting the process.

If the user passes the authentication token and service endpoint as
arguments or sets the respective environment variables, these settings
will be used. The cache will not be updated with these values.

The user can force the cached information be refreshed at any time using
the argument --refresh-cache.

The user can disable this feature by passing the argument --no-cache or
setting the environment variable FMCLIENT_NO_CACHE to True.

Caching reduced the time spent on authentication and service discovery
from ~400 ms to ~7 ms.

TEST PLAN:
PASS: CLI operations are working correctly
PASS: Cache data is correctly stored and retrieved from the keyring
PASS: After the first CLI operation, the cached information is being
      used
PASS: When the token expires, a new one is created on the next CLI
      operation
PASS: Keyring entry is removed when user logs out
PASS: The cache is not used when using the no-cache argument
PASS: the cache data is refreshed when using the refresh-cache argument
PASS: Initial install
PASS: DC-1000 scale test

Story: 2011322
Task: 51925

Change-Id: I53bac6fd268329cd6125ca27348f2f45609c0ab7
Signed-off-by: Reynaldo P Gomes <reynaldo.patronegomes@windriver.com>