Add sysinv dependency to fm-rest-api image
A recent change added the sysinv module to a
file in the cgtsclient package [1]. That caused the
fm-rest-api to fail when using the stx-fm-rest-api
docker image.

This review adds the sysinv package as a DIS_PACKAGE
in the stx-fm-rest-api dockerfile. It also adds the
dependency to the tox file.

[1] - https://review.opendev.org/c/starlingx/config/+/952590

Test Plan:
PASS - Build fm-rest-api package and stx-fm-rest-api image
PASS - Use new docker image in the STX-O deployment
PASS - STX-O successfully applied
PASS - fm-rest-api pod running normally

Closes-Bug: #2116944

Change-Id: Ib561c6989b8763aabf43c3fdc03c502d9e119014
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>